requirements.txt
langchain
langchain-core
langchain-openai
langchain-community
langgraph
langsmith
openai
faiss-cpu
python-dotenv